Finalize engine/game boundary and remove stale milestone content
Split game composition into src/game (core_game_ui target) with a
configure-time guard keeping src/core free of game-layer includes, add
the script-first content pipeline (GameContentRuntime, scripted GUI
painter, game.coreproject/ui sample) and remove stale milestone docs
(PERFORMANCE_BUDGET_0_4/0_7/0_8/0_9, RELEASE_STATUS_1_0_RC_GPU,
VIC3_JOMINI_GAP_MATRIX, VALIDATION_1_0.txt). content/base and demo/
remain as the bundled sample game; manifests regenerated for 354
first-party files.

+++ src/core/editor/MapEditorSystem.hpp
@@ -49,8 +49,10 @@ struct EditorBrushState {
 
 // A single cell modification in a paint stroke
 struct EditorCellEdit {
-    std::uint32_t page_x = 0;          // page grid coordinate
-    std::uint32_t page_y = 0;
+    // Signed: world pixel coordinates run negative west/south of the origin,
+    // so page indices do too.
+    std::int64_t page_x = 0;           // page grid coordinate
+    std::int64_t page_y = 0;
     std::uint32_t local_x = 0;         // pixel within 128x128 page
     std::uint32_t local_y = 0;
     std::uint16_t old_value = 0;       // previous encoded province ID
